# Rewriting for loops as closures provides another speed up