Don't hate me!  I appreciate your attention to this and the last PR.  I'm sure I'm just as wrong (once again), but, rewriting in the functional form *seems to* provide a hefty speed up as well.  See:

```
hyperfine -w 3 "cat ../../kjvbible_x10.txt | ./target/release/countwords" "cat ../../kjvbible_x10.txt | ../optimized/target/release/countwords"
Benchmark 1: cat ../../kjvbible_x10.txt | ./target/release/countwords
  Time (mean ± σ):     175.2 ms ±   0.5 ms    [User: 172.8 ms, System: 10.5 ms]
  Range (min … max):   174.5 ms … 176.2 ms    16 runs

Benchmark 2: cat ../../kjvbible_x10.txt | ../optimized/target/release/countwords
  Time (mean ± σ):     223.6 ms ±   0.8 ms    [User: 216.1 ms, System: 14.2 ms]
  Range (min … max):   222.3 ms … 225.6 ms    13 runs

Summary
  'cat ../../kjvbible_x10.txt | ./target/release/countwords' ran
    1.28 ± 0.01 times faster than 'cat ../../kjvbible_x10.txt | ../optimized/target/release/countwords'
```