@benhoyt 2022-03-02
> Don't hate me! 

I definitely don't. :-)

I'm also not keen to make changes to this code at this point, particularly Andrew's Rust code, because I asked him specially to write it for this article, so I'd like it to stay as his original code. I see that article (and the code) more as a "point in time" snapshot. What I'd suggest is that you make changes on a fork and write your own article explaining how you improved and sped up @BurntSushi's Rust code ... that would make a fun article. ;-)

It's certainly not you or your code. See the [note](https://github.com/benhoyt/countwords#note) in the README about how I'm not taking new submissions any more. I might make this repo read-only (archived) just to make that a bit clearer.